requirements.txt
rich
prompt_toolkit
pyyaml
psutil
requests
pyautogui
pillow
selenium
webdriver_manager
selenium_stealth
textual
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
cryptography
aiohttp
python-docx
python-pptx
lxml
opencv-python
numpy
pygetwindow
schedule
